refactor validator into two files
